requirements.txt
python-dotenv
langchain
langchain-google-genai
langchain-community
fastapi
uvicorn